requirements.txt
psycopg2==2.8.6
python-dotenv==0.19.0
requests==2.26.0
setuptools==52.0.0.post20210125